add confirmation box before deleting

+++ NotFightClub/NotFightClub_WebAPI/Angular/src/app/user/edit-profile/edit-profile.component.ts
@@ -75,10 +75,12 @@ export class EditProfileComponent implements OnInit {
     } else {
       let id1 = JSON.parse(id).userId
       console.log(id1)
-      this.userService.deleteUser(id1).subscribe(
-        id => {
-          this.router.navigate(['login']);
-        });
+      if (confirm("Are you sure you want to delete your profile?\nAll information associated to this user profile will be permanently deleted.")) {
+        this.userService.deleteUser(id1).subscribe(
+          id => {
+            this.router.navigate(['login']);
+          });
+      }
     }
   }
